fixup removed bionic i386 lxc images
This updates the vagrant container host to 2204, which then uses distrobuilder to
build the LXC containers (bionic 1804) 64 and 32 bit from LXC's templates inside
the host instead of attempting to download them. This is necessary because LXC removed
32 bit containers from their website.

This is just a bandaid for the time being. 2004 and higher do not allow 32 bit
LXC images to be built (they break during building).

Eventually we need to migrate the wine build script to build 32 bit inside 64 bit
then get rid of the LXC containers altogether and just use Vagrant.